requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.8
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.15.2
async-timeout==5.0.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorlog==6.9.0
contourpy==1.3.0
cycler==0.12.1
exceptiongroup==1.2.2
filelock==3.17.0
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.2.0
greenlet==3.2.1
idna==3.10
importlib_resources==6.5.2
iniconfig==2.0.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.7
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
multidict==6.1.0
narwhals==1.38.0
networkx==3.2.1
numpy==1.23.0
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
optuna==4.3.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.1
pluggy==1.5.0
propcache==0.3.0
psutil==7.0.0
pyparsing==3.2.1
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.7.1
seaborn==0.13.2
six==1.17.0
SQLAlchemy==2.0.40
sympy==1.13.1
threadpoolctl==3.5.0
tomli==2.2.1
torch==2.6.0+cu118
torch-geometric==2.6.1
torch_cluster==1.6.3
torch_scatter==2.1.2
torch_sparse==0.6.18
torch_spline_conv==1.2.2
torchaudio==2.6.0+cu118
torchcde==0.2.5
torchdiffeq==0.2.5
torchsde==0.2.6
torchvision==0.21.0+cu118
tqdm==4.67.1
trampoline==0.1.2
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
yarl==1.18.3
zipp==3.21.0
